添加滑动功能

+++ lrcview/src/main/java/me/wcy/lrcview/LrcEntry.java
@@ -24,7 +24,7 @@ class LrcEntry implements Comparable<LrcEntry> {
     private long time;
     private String text;
     private StaticLayout staticLayout;
-    private int offset = Integer.MIN_VALUE;
+    private float offset = Float.MIN_VALUE;
 
     private LrcEntry(long time, String text) {
         this.time = time;
@@ -50,11 +50,11 @@ class LrcEntry implements Comparable<LrcEntry> {
         return staticLayout.getHeight();
     }
 
-    public int getOffset() {
+    public float getOffset() {
         return offset;
     }
 
-    public void setOffset(int offset) {
+    public void setOffset(float offset) {
         this.offset = offset;
     }
